modified project structure. added lib "common". added RequestReader, RequestSender, ResponseSender, ResponseReader implementation. added sendData implementation method to ConnectionModule. Commands are soon to be modified (command result will RETURN value, not print). Continue dividing the programme into classes.
