+++ client/src/main/java/clientModules/connection/SendDataAble.java
@@ -1,0 +1,5 @@
+package clientModules.connection;
+
+public interface SendDataAble {
+    void sendData(byte[] data);
+}
