--- client/src/main/java/clientModules/request/requests/CommandDescriptionsRequest.java
@@ -1,5 +1,0 @@
-package clientModules.request.requests;
-
-
-public class CommandDescriptionsRequest implements Request {
-}
