--- client/src/main/java/clientModules/request/requests/Request.java
@@ -1,4 +1,0 @@
-package clientModules.request.requests;
-
-public interface Request {
-}
