--- client/src/main/java/clientModules/response/Response.java
@@ -1,4 +1,0 @@
-package clientModules.response;
-
-public interface Response {
-}
