+++ client/src/main/java/clientModules/response/manager/ResponseManager.java
@@ -1,0 +1,4 @@
+package clientModules.response.manager;
+
+public class ResponseManager {
+}
