+++ client/src/main/java/clientModules/response/reader/ResponseReadAble.java
@@ -1,0 +1,7 @@
+package clientModules.response.reader;
+
+import java.nio.ByteBuffer;
+
+public interface ResponseReadAble<T> {
+    T readResponse(ByteBuffer buffer);
+}
