+++ common/src/main/java/module-info.java
@@ -1,0 +1,7 @@
+module common {
+    requires com.fasterxml.jackson.annotation;
+    exports defaultClasses;
+    exports responses;
+    exports requests;
+    exports commands;
+}
