+++ common/src/main/java/requests/CommandDescriptionsRequest.java
@@ -1,0 +1,7 @@
+package requests;
+
+
+import java.io.Serializable;
+
+public class CommandDescriptionsRequest implements Request, Serializable {
+}
