+++ common/src/main/java/requests/Request.java
@@ -1,0 +1,4 @@
+package requests;
+
+public interface Request {
+}
