+++ common/src/main/java/responses/ExecutionResultResponse.java
@@ -1,0 +1,15 @@
+package responses;
+
+import java.io.Serializable;
+
+public class ExecutionResultResponse implements Response, Serializable {
+    private final String result;
+
+    public ExecutionResultResponse(String result) {
+        this.result = result;
+    }
+
+    public String getResult() {
+        return result;
+    }
+}
