+++ common/src/main/java/responses/Response.java
@@ -1,0 +1,4 @@
+package responses;
+
+public interface Response {
+}
