+++ server/src/main/java/commands/BaseCommand.java
@@ -1,7 +1,6 @@
 package commands;
 
 import java.io.IOException;
-import java.io.Serializable;
 
 /**
  * Abstraction class for all commands.
@@ -9,7 +8,7 @@ import java.io.Serializable;
  * @author Dmitrii Chebanenko
  */
 
-public abstract class BaseCommand implements Serializable {
+public abstract class BaseCommand {
     /**
      * Abstract command call method.
      *
