+++ server/src/main/java/serverModules/callerBack/CallerBack.java
@@ -1,0 +1,21 @@
+package serverModules.callerBack;
+
+import java.net.InetAddress;
+
+public class CallerBack {
+    private final InetAddress address;
+    private final int port;
+
+    public CallerBack(InetAddress address, int port) {
+        this.address = address;
+        this.port = port;
+    }
+
+    public InetAddress getAddress() {
+        return address;
+    }
+
+    public int getPort() {
+        return port;
+    }
+}
