--- server/src/main/java/serverModules/request/RequestReadAble.java
@@ -1,7 +1,0 @@
-package serverModules.request;
-
-import java.net.DatagramPacket;
-
-public interface RequestReadAble<T> {
-    T readRequest(DatagramPacket packet);
-}
