+++ server/src/main/java/serverModules/request/manager/RequestManager.java
@@ -1,0 +1,4 @@
+package serverModules.request.manager;
+
+public class RequestManager {
+}
