+++ server/src/main/java/serverModules/request/reader/RequestReadAble.java
@@ -1,0 +1,7 @@
+package serverModules.request.reader;
+
+import java.net.DatagramPacket;
+
+public interface RequestReadAble<T> {
+    T readRequest(DatagramPacket packet);
+}
