--- server/src/main/java/serverModules/request/requests/CommandDescriptionsRequest.java
@@ -1,5 +1,0 @@
-package serverModules.request.requests;
-
-
-public class CommandDescriptionsRequest implements Request {
-}
