--- server/src/main/java/serverModules/request/requests/Request.java
@@ -1,4 +1,0 @@
-package serverModules.request.requests;
-
-public interface Request {
-}
