--- server/src/main/java/serverModules/response/responses/ExecutionResultResponse.java
@@ -1,13 +1,0 @@
-package serverModules.response.responses;
-
-public class ExecutionResultResponse implements Response {
-    private final String result;
-
-    public ExecutionResultResponse(String result) {
-        this.result = result;
-    }
-
-    public String getResult() {
-        return result;
-    }
-}
