--- server/src/main/java/serverModules/response/responses/Response.java
@@ -1,4 +1,0 @@
-package serverModules.response.responses;
-
-public interface Response {
-}
