+++ server/src/main/java/serverModules/response/sender/ResponseAble.java
@@ -1,7 +1,9 @@
 package serverModules.response.sender;
 
-import java.net.DatagramSocket;
+import serverModules.connection.ConnectionModule;
+
+import java.net.InetAddress;
 
 public interface ResponseAble<T> {
-    void sendResponse(DatagramSocket socket, String address, int port, T obj);
+    void sendResponse(ConnectionModule module, InetAddress address, int port, T obj);
 }
